requirements.txt
flask
flask-cors
flask-jwt-extended
Flask-Migrate
flask-smorest
Flask-SQLAlchemy
gunicorn
marshmallow
python-dotenv
